pyproject.toml
[tool.poetry]
name = "grocy-rest-api-client"
version = "4.2.0"
description = "A client library for accessing Grocy REST API"
authors = []
readme = "README.md"
packages = [
    {include = "grocy_rest_api_client"},
]
include = ["CHANGELOG.md", "grocy_rest_api_client/py.typed"]


[tool.poetry.dependencies]
python = "^3.11"
httpx = ">=0.20.0,<0.28.0"
attrs = ">=21.3.0"
python-dateutil = "^2.8.0"
openfoodfacts = "^0.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["F", "I", "UP"]
